All auth for api and csv to transaction

+++ app/Http/Controllers/TransactionController.php
@@ -2,23 +2,29 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\DeleteTransactionRequest;
+use App\Http\Requests\ShowTransactionRequest;
 use App\Http\Requests\StoreTransactionRequest;
 use App\Http\Requests\UpdateTransactionRequest;
-use App\Http\Resources\TransactionResource;
+use App\Http\Resources\MapResource;
 use App\Models\Transaction;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Validation\UnauthorizedException;
 
 class TransactionController extends Controller
 {
+    private const EXCEPTIONS_MESSAGE_UNAUTHORIZED = 'exceptions.message.unauthorized';
+
     /**
      * Display a listing of the resource.
      *
-     * @return TransactionResource
+     * @param ShowTransactionRequest $request
+     * @return MapResource
      */
-    public function index()
+    public function index(ShowTransactionRequest $request)
     {
-        return Transaction::latest()
-            ->filter(request(['date', 'description', 'file', 'type', 'bank']))
+        return $request->user()->transactions()->latest()
+            ->filter(request(['date', 'description', 'file', 'type', 'bank', 'start', 'until', 'expenses', 'incomes']))
             ->paginate(18)
             ->withQueryString();
     }
@@ -31,13 +37,13 @@ class TransactionController extends Controller
      */
     public function store(StoreTransactionRequest $request)
     {
-        $transaction = (new Transaction)->create([
+        $transaction = $request->user()->transactions()->create([
             'date' => $request->get('date'),
             'amount' => $request->get('amount'),
             'description' => $request->get('description'),
             'file' => $request->get('file'),
             'type' => $request->get('type'),
-            'bank' => $request->get('bank'),
+            'bank' => $request->get('bank')
         ]);
 
         return response()->json(["transaction" => $transaction], 200);
@@ -46,12 +52,17 @@ class TransactionController extends Controller
     /**
      * Display the specified resource.
      *
+     * @param ShowTransactionRequest $request
      * @param Transaction $transaction
      * @return JsonResponse
      */
-    public function show(Transaction $transaction)
+    public function show(ShowTransactionRequest $request, Transaction $transaction)
     {
-        return response()->json(["transaction" => $transaction], 200);
+        if($request->user()->id === $transaction->user()->first()->id) {
+            return response()->json(["transaction" => $transaction], 200);
+        }
+
+        throw new UnauthorizedException(config(self::EXCEPTIONS_MESSAGE_UNAUTHORIZED));
     }
 
     /**
@@ -63,27 +74,36 @@ class TransactionController extends Controller
      */
     public function update(UpdateTransactionRequest $request, Transaction $transaction)
     {
-        $transaction->update([
-            'date' => $request->get('date') ?? $transaction->date,
-            'amount' => $request->get('amount') ?? $transaction->amount,
-            'description' => $request->get('description') ?? $transaction->description,
-            'file' => $request->get('file') ?? $transaction->file,
-            'type' => $request->get('type') ?? $transaction->type,
-            'bank' => $request->get('bank') ?? $transaction->bank,
-        ]);
+        if($request->user()->id === $transaction->user()->first()->id) {
+            $transaction->update([
+                'date' => $request->get('date') ?? $transaction->date,
+                'amount' => $request->get('amount') ?? $transaction->amount,
+                'description' => $request->get('description') ?? $transaction->description,
+                'file' => $request->get('file') ?? $transaction->file,
+                'type' => $request->get('type') ?? $transaction->type,
+                'bank' => $request->get('bank') ?? $transaction->bank,
+            ]);
 
-        return response()->json(["transaction" => $transaction], 200);
+            return response()->json(["transaction" => $transaction], 200);
+        }
+
+        throw new UnauthorizedException(config(self::EXCEPTIONS_MESSAGE_UNAUTHORIZED));
     }
 
     /**
      * Remove the specified resource from storage.
      *
+     * @param DeleteTransactionRequest $request
      * @param Transaction $transaction
      * @return JsonResponse
      */
-    public function destroy(Transaction $transaction)
+    public function destroy(DeleteTransactionRequest $request, Transaction $transaction)
     {
-        $transaction->destroy($transaction->id);
-        return response()->json(["success"], 200);
+        if($request->user()->id === $transaction->user()->first()->id) {
+            $transaction->destroy($transaction->id);
+            return response()->json(["success"], 200);
+        }
+
+        throw new UnauthorizedException(config(self::EXCEPTIONS_MESSAGE_UNAUTHORIZED));
     }
 }
